Moved TradManager to BL library.
