+++ moddingSuite.BL/Utils/StdUtils.cs
@@ -8,6 +8,21 @@ namespace moddingSuite.BL.Utils
 {
     public static class StdUtils
     {
+        public static string ByteArrayToBigEndianHexByteString(byte[] data)
+        {
+            if (data == null)
+                return string.Empty;
+
+            var stringBuilderb = new StringBuilder();
+
+            stringBuilderb.Append(string.Empty);
+
+            foreach (var b in data)
+                stringBuilderb.Append(string.Format("{0:X2}", b));
+
+            return stringBuilderb.ToString();
+        }
+
         public static T ByteArrayToStructure<T>(byte[] bytes) where T : struct
         {
             GCHandle handle = GCHandle.Alloc(bytes, GCHandleType.Pinned);
@@ -76,6 +91,34 @@ namespace moddingSuite.BL.Utils
             Marshal.FreeHGlobal(ptr);
 
             return arr;
+        }
+
+        public static byte[] CreateLocalisationHash(string text, int maxSize = 8)
+        {
+            long hash = 0;
+            for (int i = 0; i < maxSize; ++i)
+            {
+                int value;
+                ushort chr = text[i];
+
+                if (chr == 0)
+                    break;
+
+                if ('0' <= chr && chr <= '9')
+                    value = 1 + chr - '0';
+                else if ('A' <= chr && chr <= 'Z')
+                    value = 2 + '9' - '0' + chr - 'A';
+                else if (chr == '_')
+                    value = 3 + '9' - '0' + 'Z' - 'A';
+                else if ('a' <= chr && chr <= 'z')
+                    value = 4 + '9' - '0' + 'Z' - 'A' + chr - 'a';
+                else
+                    throw new InvalidDataException("");
+
+                hash = (hash << 6) | value;
+            }
+
+            return BitConverter.GetBytes(hash);
         }
     }
 }
