+++ moddingSuite/Model/Ndfbin/Types/AllTypes/NdfHash.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using moddingSuite.Util;
+using moddingSuite.BL.Utils;
 
 namespace moddingSuite.Model.Ndfbin.Types.AllTypes
 {
@@ -36,7 +36,7 @@ namespace moddingSuite.Model.Ndfbin.Types.AllTypes
 
         public override string ToString()
         {
-            return Utils.ByteArrayToBigEndianHexByteString(Value);
+            return StdUtils.ByteArrayToBigEndianHexByteString(Value);
         }
     }
 }
