+++ moddingSuite/View/Extension/LocalisationHashValueConverter.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Globalization;
 using System.Windows.Data;
+using moddingSuite.BL.Utils;
 using moddingSuite.Util;
 
 namespace moddingSuite.View.Extension
@@ -11,7 +12,7 @@ namespace moddingSuite.View.Extension
 
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return Utils.ByteArrayToBigEndianHexByteString((byte[]) value);
+            return StdUtils.ByteArrayToBigEndianHexByteString((byte[]) value);
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
